+++ backend/src/activity_log/activity_utils.rs
