+++ backend/src/activity_log/mod.rs
@@ -1,3 +1,3 @@
-pub mod activity_log;
 pub mod activity_serializer;
+pub mod activity_utils;
 pub mod package_add_activity;
