+++ backend/src/api/activity.rs
@@ -1,4 +1,4 @@
-use crate::activity_log::activity_log::{Activity, ActivityLog};
+use crate::activity_log::activity_utils::{Activity, ActivityLog};
 use crate::api::models::authenticated::Authenticated;
 use rocket::response::status::NotFound;
 use rocket::serde::json::Json;
